requirements.txt
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
requests>=2.31.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
flask>=2.3.0
flask-cors>=4.0.0